consolidate constants
update routing and apply better comments
