+++ internal/router/registry.go
@@ -43,7 +43,7 @@ func (r *RouteRegistry) RegisterWithMethod(route string, handler http.HandlerFun
 
 func (r *RouteRegistry) RegisterProxyRoute(route string, handler http.HandlerFunc, description string, method string) {
 	wrappedHandler := func(w http.ResponseWriter, req *http.Request) {
-		ctx := context.WithValue(req.Context(), constants.ProxyPathPrefix, route)
+		ctx := context.WithValue(req.Context(), constants.ContextRoutePrefixKey, route)
 		handler(w, req.WithContext(ctx))
 	}
 	r.registerWithMethod(route, wrappedHandler, description, method, true)
